Add files via upload
This machine learning model is trained for predicting the crop label that is most likely to result in maximum productivity based on those features. By providing values for these features as input to the model, it will output the predicted probabilities for each crop class. The crop with the highest predicted probability is considered the best choice for maximizing productivity under the given conditions. This prediction is based on the patterns learned by the model during training from the provided dataset.This model is further integrated with digital twin model for increasing the overall efficiency.